pyproject.toml
[project]
name = "cost-observability-control"
version = "1.0.0"
description = "Cost Observability & Control - Databricks compute consumption and spend analytics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "databricks-sdk>=0.38.0",
    "databricks-sql-connector>=3.6.0,<4.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "cachetools>=7.0.0",
    "python-multipart>=0.0.9",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "psycopg[binary,pool]>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
